refactor(types): drop stale type-ignore comments and tighten annotations
Remove # type: ignore directives whose underlying issues were fixed
upstream or by previous casts. Where mypy strict still flagged a real
no-any-return at a third-party boundary (json/numpy/soxr/HF processor /
ASGI scope / DuckDB / scipy), bind the result to an explicit-typed
local before returning so the boundary becomes visible to readers and
to the type checker.

+++ asrbench/engine/search/significance.py
@@ -208,7 +208,7 @@ def suggest_epsilon(
     """
     n_words = max(1.0, dataset_duration_s * wpm / 60.0)
     se = (base_wer * (1.0 - base_wer) / n_words) ** 0.5
-    return round(safety * se, 3)
+    return float(round(safety * se, 3))
 
 
 def sensitivity_score(
